ugc_flask/etl/requirements.txt
aiochclient==2.6.0
aiokafka==0.11.0
async-timeout==4.0.3
packaging==24.1
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.1
typing_extensions==4.12.2
tzlocal==5.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
frozenlist==1.4.1
idna==3.10
multidict==6.1.0
yarl==1.11.1